+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 0.4.0 - 2020-03-09
+### Changed:
+ - TestCase is now abstract 
+### Added:
+ - assertPromiseFulfillsWithInstanceOf() to check class of the resolution value 
+
 ## 0.3.0 - 2020-03-08
 ### Updated:
  - Dependencies